chore: Renamed Core/RichServerResponse.mts to Core/rich-server-response.mts.
