+++ packages/architectura/src/Core/rich-server-response.mts
