+++ packages/architectura/src/definition/interface/execution-context-instantiation.interface.mts
@@ -1,6 +1,7 @@
-import type { RichServerResponse } from "../../Core/RichServerResponse.mjs";
-
 import type { RichClientRequest } from "../../Core/rich-client-request.mjs";
+
+import type { RichServerResponse } from "../../Core/rich-server-response.mjs";
+
 
 
 interface ExecutionContextInstantiationInterface
